% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLR.R
\name{contours}
\alias{contours}
\title{Produce level curve for a KLR object.}
\usage{
contours(KLRobj, dims = 1:2, res = 100, levels = c(0.5))
}
\arguments{
\item{KLRobj}{An object of class \code{KLR}.}

\item{dims}{Dimensions for which t0 produce contours. Other dimensions are set the mean.}

\item{res}{Resolution of the grid.}

\item{levels}{Levels at which to produce level curves.}
}
\value{
A list containing the desired curves. Each list has a \code{level} attribute stating the respective level as well as \code{x} and \code{y} attributes defining the curve.
}
\description{
Produce level curve for a KLR object.
}
