% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv.KLR}
\alias{cv.KLR}
\title{Cross-validation on Kernel Logistic Regression}
\usage{
cv.KLR(
  y,
  x,
  n_folds = 5,
  kernel = c("gaussian", "polynomial")[1],
  lambda = c(1, 0.1, 0.01, 0.001),
  sigma2 = c(5, 2, 1, 0.5),
  d = 3,
  threshold = 1e-06,
  max_iter = 1e+05
)
}
\arguments{
\item{y}{A \code{n x 1} column vector containing the responses (0-1).}

\item{x}{A \code{n x p} matrix containing the covariates.}

\item{n_folds}{Number of folds in the CV.}

\item{kernel}{The kernel to use. Either \code{gaussian} (default) or \code{polynomial}.}

\item{lambda}{The regularization parameter(s).}

\item{sigma2}{The scale(s) in the \code{gaussian} and \code{polynomial} kernel. See details.}

\item{d}{The degree(s) in the \code{polynomial} kernel.}

\item{threshold}{The convergence threshold.}

\item{max_iter}{The maximum number of iterations.}
}
\value{
A list containing:
\describe{
\item{\code{mpe}}{The mean prediction error across all folds for all combinations of parameters.}
\item{\code{lambda_min}}{The selected value of \code{lambda}.}
\item{\code{sigma2_min}}{The selected value of \code{sigma2}.}
\item{\code{d_min}}{The selected value of \code{d}.}
}
}
\description{
This function performs cross-validation to select values of tuning parameters.
}
\details{
The \code{gaussian} kernel has the following form:
\deqn{exp(-||x-y||^2/sigma2).}
The \code{polynomial} kernel has the following form:
\deqn{(1+x'y/sigma2)^d.}
}
\seealso{
\link{KLR}
}
