\name{KLR-package}
\alias{KLR-package}
\alias{KLR}
\docType{package}
\title{
\packageTitle{KLR}
}
\description{
\packageDescription{KLR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{KLR}
\packageIndices{KLR}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{KLR}

Maintainer: \packageMaintainer{KLR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
