% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLR.R
\name{predict.KLR}
\alias{predict.KLR}
\title{Predict using a KLR fit}
\usage{
\method{predict}{KLR}(object, newx = object$x, ...)
}
\arguments{
\item{object}{An object of class \code{KLR}.}

\item{newx}{The \code{m x p} matrix of observations at which to perform prediction.}

\item{...}{Extra arguments (not used).}
}
\value{
The \code{m x 1} vector of predicted probabilities.
}
\description{
Prediction
}
\seealso{
\link{KLR}
}
