% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLR.R
\name{KLR}
\alias{KLR}
\title{Kernel Logistic Regression}
\usage{
KLR(y, x, kernel = c("gaussian", "polynomial")[1], lambda = 0.01,
  sigma2 = 1, d = 3, threshold = 1e-06, max_iter = 1e+05)
}
\arguments{
\item{y}{A \code{n x 1} column vector containing the responses (0-1).}

\item{x}{A \code{n x p} matrix containing the covariates.}

\item{kernel}{The kernel to use. Either \code{gaussian} (default) or \code{polynomial}.}

\item{lambda}{The regularization parameter.}

\item{sigma2}{The scale in the \code{gaussian} and \code{polynomial} kernel. See details.}

\item{d}{The degree in the \code{polynomial} kernel.}

\item{threshold}{The convergence threshold.}

\item{max_iter}{The maximum number of iterations.}
}
\description{
This function does
}
