% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLR.R
\name{KLR}
\alias{KLR}
\title{Kernel Logistic Regression}
\usage{
KLR(y, x, kernel = c("gaussian", "polynomial")[1], lambda = 0.01,
  sigma2 = 1, d = 3, threshold = 1e-06, max_iter = 1e+05)
}
\arguments{
\item{y}{A \code{n x 1} column vector containing the responses (0-1).}

\item{x}{A \code{n x p} matrix containing the covariates.}

\item{kernel}{The kernel to use. Either \code{gaussian} (default) or \code{polynomial}.}

\item{lambda}{The regularization parameter.}

\item{sigma2}{The scale in the \code{gaussian} and \code{polynomial} kernel. See details.}

\item{d}{The degree in the \code{polynomial} kernel.}

\item{threshold}{The convergence threshold.}

\item{max_iter}{The maximum number of iterations.}
}
\value{
A list containing:
\describe{
\item{\code{x}}{The original \code{x}.}
\item{\code{alpha}}{The vector of fitted weights.}
\item{\code{kernel}}{The kernel.}
\item{\code{sigma2}}{The scale parameter.}
\item{\code{d}}{The polynomial degree.}
}
}
\description{
This function fit a kernel logistic regression model to the data (\code{y}, \code{x}) using some pre-specified kernel. The return list contains the estimated kernel weights as well as the original data to perform predictions.
}
\details{
The \code{gaussian} kernel has the following form:
\deqn{exp(-||x-y||^2/sigma2).}
The \code{polynomial} kernel has the following form:
\deqn{(1+x'y/sigma2)^d.}
}
\seealso{
\link{predict.KLR}
}
